# internal/api/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flagcel API
  description: Feature flag service with CEL-based targeting rules.
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
...
paths:
...
  /flags:
    get:
      tags: [Flags]
      summary: List flags
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Flags list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagListEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Flags]
      summary: Create or replace a flag
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlagRequest"
      responses:
        "200":
          description: Flag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
  /flags/{key}:
    parameters:
      - $ref: "#/components/parameters/FlagKey"
    get:
      tags: [Flags]
      summary: Get a flag by key
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Flags]
      summary: Delete a flag and all of its rules
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Flag deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /environments/{environment_id}/flags:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
    get:
      tags: [Flags]
      summary: List flags in an environment
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Flags list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagListEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/EnvironmentNotFound" }
    post:
      tags: [Flags]
      summary: Create or replace a flag in an environment
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlagRequest"
      responses:
        "200":
          description: Flag created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/EnvironmentNotFound" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
  /environments/{environment_id}/flags/{key}:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/FlagKey"
    get:
      tags: [Flags]
      summary: Get a flag by key in an environment
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
    delete:
      tags: [Flags]
      summary: Delete a flag and all of its rules in an environment
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Flag deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
  /environments/{environment_id}/flags/{key}/evaluate:
    parameters:
      - $ref: "#/components/parameters/EnvironmentID"
      - $ref: "#/components/parameters/FlagKey"
    post:
      tags: [Flags]
      summary: Evaluate one flag with diagnostic trace
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvalRequest"
      responses:
        "200":
          description: Evaluation trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalTraceEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
        "422": { $ref: "#/components/responses/InvalidRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /eval:
    post:
      tags: [Evaluation]
      summary: Evaluate all flags
      security:
        - bearerAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvalRequest"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalAllEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /eval/definitions:
    get:
      tags: [Evaluation]
      summary: Get raw flag definitions for SDK polling
      security:
        - bearerAPIKey: []
      parameters:
        - in: header
          name: If-None-Match
          schema:
            type: string
          description: Previous ETag returned by this endpoint.
      responses:
        "200":
          description: Flag definitions and referenced context schemas
          headers:
            ETag:
              schema:
                type: string
              description: Current in-memory definitions version.
            Cache-Control:
              schema:
                type: string
              description: Cache validator policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalDefinitionsEnvelope"
        "304":
          description: Definitions have not changed since the supplied ETag
          headers:
            ETag:
              schema:
                type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
  /eval/{key}:
    parameters:
      - $ref: "#/components/parameters/FlagKey"
    post:
      tags: [Evaluation]
      summary: Evaluate one flag
      security:
        - bearerAPIKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvalRequest"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalEnvelope"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/FlagNotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: flagcel_session
    bearerAPIKey:
      type: http
      scheme: bearer
  parameters:
    EnvironmentID:
      name: environment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FlagKey:
      name: key
      in: path
      required: true
      schema:
        type: string
        example: new-checkout
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [message, data]
      properties:
        message:
          type: string
          example: success
        data: {}
...
    Rollout:
      type: object
      required: [percentage]
      properties:
        percentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 25
        bucket_by:
          type: string
          example: user.id
    Rule:
      type: object
      required: [id, expression, rollout, value, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        description: { type: string }
        expression:
          type: string
          example: "user.country == 'US'"
        rollout:
          $ref: "#/components/schemas/Rollout"
        value:
          $ref: "#/components/schemas/FlagValue"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, nullable: true }
        deleted_by: { type: string, format: uuid, nullable: true }
    Flag:
      type: object
      required: [key, type, enabled, rules, default_value, created_at, updated_at]
      properties:
        key: { type: string, example: new-checkout }
        description: { type: string }
        type:
          type: string
          enum: [boolean, string, number, json]
        enabled: { type: boolean }
        rules:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
        default_value:
          $ref: "#/components/schemas/FlagValue"
        context_id:
          type: string
          format: uuid
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, nullable: true }
        deleted_by: { type: string, format: uuid, nullable: true }
    CreateFlagRequest:
      type: object
      required: [key]
      properties:
        key: { type: string }
        description: { type: string }
        type:
          type: string
          enum: [boolean, string, number, json]
          default: boolean
        enabled: { type: boolean, default: false }
        rules:
          type: array
          items:
            $ref: "#/components/schemas/CreateRuleRequest"
          default: []
        default_value:
          allOf:
            - $ref: "#/components/schemas/FlagValue"
          default: false
        context_id:
          type: string
          format: uuid
          nullable: true
    CreateRuleRequest:
      type: object
      required: [expression, rollout]
      properties:
        description: { type: string }
        expression: { type: string }
        rollout:
          $ref: "#/components/schemas/Rollout"
        value:
          allOf:
            - $ref: "#/components/schemas/FlagValue"
          default: true
...
    ContextField:
      type: object
      required: [path, type]
      properties:
        path: { type: string, example: user.country }
        type:
          type: string
          enum: [string, int, double, bool, timestamp, list, map]
    Context:
      type: object
      required: [id, name, fields, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ContextField"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid, nullable: true }
        deleted_by: { type: string, format: uuid, nullable: true }
...
    EvalRequest:
      type: object
      required: [context]
      properties:
        context:
          type: object
          additionalProperties: true
    EvalResponse:
      type: object
      required: [key, value_type, value]
      properties:
        key: { type: string }
        value_type:
          type: string
          enum: [boolean, string, number, json]
        value:
          $ref: "#/components/schemas/FlagValue"
    EvalMatchedRule:
      type: object
      required: [id, index, expression, value]
      properties:
        id: { type: string, format: uuid }
        index: { type: integer, minimum: 0 }
        expression: { type: string }
        value:
          $ref: "#/components/schemas/FlagValue"
    EvalRuleResult:
      type: object
      required: [id, index, expression, value, matched]
      properties:
        id: { type: string, format: uuid }
        index: { type: integer, minimum: 0 }
        expression: { type: string }
        value:
          $ref: "#/components/schemas/FlagValue"
        matched: { type: boolean }
        error: { type: string }
    EvalBucket:
      type: object
      required: [bucket_by, percentage, in_rollout, missing]
      properties:
        bucket_by: { type: string }
        bucket_value: { type: string }
        bucket_number: { type: integer, minimum: 0, maximum: 99 }
        percentage: { type: integer, minimum: 0, maximum: 100 }
        in_rollout: { type: boolean }
        missing: { type: boolean }
    EvalTraceResponse:
      type: object
      required: [key, value_type, enabled, default_value, value, reason, rule_results]
      properties:
        key: { type: string }
        value_type:
          type: string
          enum: [boolean, string, number, json]
        enabled: { type: boolean }
        default_value:
          $ref: "#/components/schemas/FlagValue"
        value:
          $ref: "#/components/schemas/FlagValue"
        reason:
          type: string
          enum: [disabled, matched_rule, default_no_match, cel_error]
        error: { type: string }
        matched_rule:
          $ref: "#/components/schemas/EvalMatchedRule"
        bucket:
          $ref: "#/components/schemas/EvalBucket"
        rule_results:
          type: array
          items:
            $ref: "#/components/schemas/EvalRuleResult"
    EvalAllResponse:
      type: object
      required: [flags]
      properties:
        flags:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/EvalFlagValue"
    EvalDefinitions:
      type: object
      required: [flags]
      properties:
        flags:
          type: array
          items:
            $ref: "#/components/schemas/FlagDefinition"
        contexts:
          type: array
          items:
            $ref: "#/components/schemas/Context"
    FlagDefinition:
      allOf:
        - $ref: "#/components/schemas/Flag"
        - type: object
          properties:
            context_schema:
              $ref: "#/components/schemas/Context"
    EvalFlagValue:
      type: object
      required: [value_type, value]
      properties:
        value_type:
          type: string
          enum: [boolean, string, number, json]
        value:
          $ref: "#/components/schemas/FlagValue"
    FlagValue:
      nullable: true
      oneOf:
        - type: boolean
        - type: string
        - type: number
        - type: object
          additionalProperties: true
        - type: array
          items: {}
...
    FlagEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Flag"
    FlagListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Flag"
...
    EvalEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/EvalResponse"
    EvalTraceEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/EvalTraceResponse"
    EvalAllEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/EvalAllResponse"
    EvalDefinitionsEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/EvalDefinitions"
...
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - FLAG_NOT_FOUND
            - RULE_NOT_FOUND
            - ENVIRONMENT_NOT_FOUND
            - ENVIRONMENT_KEY_TAKEN
            - ENVIRONMENT_IN_USE
            - DEFAULT_ENVIRONMENT
            - CONTEXT_NOT_FOUND
            - CONTEXT_NAME_TAKEN
            - API_KEY_NOT_FOUND
            - INVALID_REQUEST
            - BAD_REQUEST
            - AUTH_NOT_CONFIGURED
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - FORBIDDEN
            - INTERNAL_ERROR
        message:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    FlagNotFound:
      description: Flag not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    EnvironmentNotFound:
      description: Environment not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidRequest:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"